Update tweets.py
This code is a Python program that retrieves a specified number of tweets from a given Twitter username and writes them into an Excel file. It uses the snscrape and pandas libraries to add the tweets into a DataFrame and write them into an Excel file. It also removes the timezone information from the dates of the tweets. The code has a robust structure that can handle various errors and provides debugging messages. Additionally, it has additional features such as retrieving data from multiple accounts and setting the name of the Excel file based on the account names automatically.